Move QType from subdivision to quadedge
Move QType from subdivision package to the quadedge package
